--- alcon/alcon/New Group/MainTabBarController.swift
@@ -1,29 +1,0 @@
-//
-//  MainTabBarController.swift
-//  alcon
-//
-//  Created by 工藤海斗 on 2021/03/26.
-//
-
-import UIKit
-
-class MainTabBarController: UITabBarController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
