+++ alcon/alcon/ViewControllers/CalendarAndDrinkingRecordViewController.swift
@@ -1,0 +1,18 @@
+//
+//  CalendarAndDrinkingRecordViewController.swift
+//  alcon
+//
+//  Created by 工藤海斗 on 2021/03/26.
+//
+
+import UIKit
+
+class CalendarAndDrinkingRecordViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .blue
+        title = "カレンダー"
+        print("CalendarAndDrinkingRecordViewController")
+    }
+}
