+++ alcon/alcon/ViewControllers/SettingsViewController.swift
@@ -1,0 +1,18 @@
+//
+//  SettingsViewController.swift
+//  alcon
+//
+//  Created by 工藤海斗 on 2021/03/26.
+//
+
+import UIKit
+
+class SettingsViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .yellow
+        title = "設定"
+        print("SettingsViewController")
+    }
+}
